scripts/requirements.txt
supabase>=2.0.0
python-dotenv>=1.0.0
faker>=20.0.0
requests>=2.31.0
beautifulsoup4>=4.12.0
